code/deployment/app/requirements.txt
streamlit
numpy
pillow
requests